Refactor ChartEditModal tabs
