+++ components/charts/ChartEditModal.tsx
@@ -1,1472 +1,94 @@
 "use client"
 
-import React, { useState, useRef, useEffect } from "react"
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
+import React, { useState } from "react"
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog"
 import { Button } from "@/components/ui/button"
 import { useAnalysisStore } from "@/stores/useAnalysisStore"
-import { X, Settings, Plus, ChevronDown, Search } from "lucide-react"
-import { Checkbox } from "@/components/ui/checkbox"
-import { Badge } from "@/components/ui/badge"
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover"
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table"
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select"
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from "@/components/ui/collapsible"
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from "@/components/ui/tooltip"
-import { ManualEntryDialog } from "../ManualEntryDialog"
-import { TriggerSignalDialog } from "../search/TriggerSignalDialog"
-import { useManualEntry } from "@/hooks/useManualEntry"
-import { EventInfo, MarkerType, LineStyle } from "@/types"
-
+import { DataSourceTab } from "./EditModal/DataSourceTab"
+import { ParametersTab } from "./EditModal/ParametersTab"
+import { AppearanceTab } from "./EditModal/AppearanceTab"
+import { EventInfo } from "@/types"
 
 export function ChartEditModal() {
   const { editingChart, editModalOpen, setEditingChart, setEditModalOpen } = useAnalysisStore()
   const [activeTab, setActiveTab] = useState<"parameters" | "datasource" | "appearance">("datasource")
-  const [lastAddedParamIndex, setLastAddedParamIndex] = useState<number | null>(null)
-  const parameterInputRefs = useRef<(HTMLInputElement | null)[]>([])
-  
-  // Event data state
-  const [events, setEvents] = useState<EventInfo[]>([
-    { id: "1", plant: "Plant A", machineNo: "M001", label: "Maintenance", labelDescription: "Regular check", event: "Scheduled Stop", eventDetail: "Monthly maintenance", start: "2024-01-15T10:00:00", end: "2024-01-15T12:00:00" },
-    { id: "2", plant: "Plant A", machineNo: "M002", label: "Production", labelDescription: "Normal run", event: "Normal Operation", eventDetail: "Batch processing", start: "2024-01-15T08:00:00", end: "2024-01-15T16:00:00" },
-    { id: "3", plant: "Plant B", machineNo: "M003", label: "Alert", labelDescription: "Warning state", event: "Temperature Warning", eventDetail: "Above threshold", start: "2024-01-15T14:30:00", end: "2024-01-15T14:45:00" },
-  ])
-  const [selectedEventIds, setSelectedEventIds] = useState<Set<string>>(new Set())
   const [selectedDataSourceItems, setSelectedDataSourceItems] = useState<EventInfo[]>([])
-  
-  // Manual entry management
-  const manualEntry = useManualEntry()
-  
-  // Trigger signal dialog state
-  const [triggerSignalDialogOpen, setTriggerSignalDialogOpen] = useState(false)
-  
-  // Search state
-  const [eventSearchTerm, setEventSearchTerm] = useState("")
-  
-  // Time offset state
-  const [startOffset, setStartOffset] = useState(0)
-  const [startOffsetUnit, setStartOffsetUnit] = useState<'min' | 'sec'>('min')
-  const [endOffset, setEndOffset] = useState(0)
-  const [endOffsetUnit, setEndOffsetUnit] = useState<'min' | 'sec'>('min')
-  const [offsetSectionOpen, setOffsetSectionOpen] = useState(false)
-
-  const handleSaveManualEntry = (data: any, editingItemId: string | null) => {
-    if (editingItemId) {
-      // Update existing entry (only in Selected Data Source, not in Event Information)
-      // Parse legend back to label and labelDescription if editing
-      const updatedItem = { ...data }
-      if (data.legend) {
-        const legendMatch = data.legend.match(/^(.+?)\s*\((.+)\)$/)
-        if (legendMatch) {
-          updatedItem.label = legendMatch[1].trim()
-          updatedItem.labelDescription = legendMatch[2].trim()
-        } else {
-          updatedItem.label = data.legend
-          updatedItem.labelDescription = ""
-        }
-      }
-      setSelectedDataSourceItems(
-        selectedDataSourceItems.map(item => 
-          item.id === editingItemId ? updatedItem : item
-        )
-      )
-    } else {
-      // Add new entry
-      const newEntry: EventInfo = {
-        ...data,
-        id: Date.now().toString()
-      }
-      
-      // Add to selected data source items
-      setSelectedDataSourceItems([...selectedDataSourceItems, newEntry])
-    }
-    
-    manualEntry.close()
-  }
-
-  const handleAddTriggerSignalResults = (results: EventInfo[]) => {
-    setSelectedDataSourceItems([...selectedDataSourceItems, ...results])
-  }
-
-  const resetStartOffset = () => {
-    setStartOffset(0)
-  }
-
-  const resetEndOffset = () => {
-    setEndOffset(0)
-  }
-
-  useEffect(() => {
-    if (lastAddedParamIndex !== null && parameterInputRefs.current[lastAddedParamIndex]) {
-      const inputElement = parameterInputRefs.current[lastAddedParamIndex]
-      inputElement?.focus()
-      inputElement?.scrollIntoView({ behavior: 'smooth', block: 'nearest' })
-      setLastAddedParamIndex(null)
-    }
-  }, [lastAddedParamIndex, editingChart?.yAxisParams?.length])
 
   if (!editingChart) return null
 
   return (
     <>
       <Dialog open={editModalOpen} onOpenChange={setEditModalOpen}>
-      <DialogContent className="max-w-7xl w-[90vw] h-[90vh] flex flex-col overflow-hidden" hideCloseButton>
-        <DialogHeader className="flex-shrink-0">
-          <div className="flex justify-between items-center">
-            <DialogTitle>Edit Chart: {editingChart.title}</DialogTitle>
-            <div className="flex gap-2">
-              <Button variant="outline" onClick={() => setEditModalOpen(false)}>
-                Cancel
-              </Button>
-              <Button
-                onClick={() => {
-                  // In a real app, you would save the changes here
-                  setEditModalOpen(false)
-                }}
-              >
-                Save Changes
-              </Button>
+        <DialogContent className="max-w-7xl w-[90vw] h-[90vh] flex flex-col overflow-hidden" hideCloseButton>
+          <DialogHeader className="flex-shrink-0">
+            <div className="flex justify-between items-center">
+              <DialogTitle>Edit Chart: {editingChart.title}</DialogTitle>
+              <div className="flex gap-2">
+                <Button variant="outline" onClick={() => setEditModalOpen(false)}>
+                  Cancel
+                </Button>
+                <Button onClick={() => setEditModalOpen(false)}>Save Changes</Button>
+              </div>
+            </div>
+          </DialogHeader>
+
+          <div className="grid grid-cols-2 gap-4 flex-1 min-h-0">
+            <div className="border rounded-lg p-4 overflow-hidden h-full flex flex-col">
+              <div className="flex gap-2 mb-4">
+                <button
+                  className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
+                    activeTab === "datasource" ? "bg-primary text-primary-foreground" : "bg-muted hover:bg-muted/80"
+                  }`}
+                  onClick={() => setActiveTab("datasource")}
+                >
+                  DataSource
+                </button>
+                <button
+                  className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
+                    activeTab === "parameters" ? "bg-primary text-primary-foreground" : "bg-muted hover:bg-muted/80"
+                  }`}
+                  onClick={() => setActiveTab("parameters")}
+                >
+                  Parameters
+                </button>
+                <button
+                  className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
+                    activeTab === "appearance" ? "bg-primary text-primary-foreground" : "bg-muted hover:bg-muted/80"
+                  }`}
+                  onClick={() => setActiveTab("appearance")}
+                >
+                  Appearance
+                </button>
+              </div>
+
+              <div className="flex-1 overflow-y-auto">
+                {activeTab === "datasource" && (
+                  <DataSourceTab
+                    selectedDataSourceItems={selectedDataSourceItems}
+                    setSelectedDataSourceItems={setSelectedDataSourceItems}
+                  />
+                )}
+                {activeTab === "parameters" && (
+                  <ParametersTab editingChart={editingChart} setEditingChart={setEditingChart} />
+                )}
+                {activeTab === "appearance" && (
+                  <AppearanceTab
+                    editingChart={editingChart}
+                    setEditingChart={setEditingChart}
+                    selectedDataSourceItems={selectedDataSourceItems}
+                  />
+                )}
+              </div>
+            </div>
+
+            <div className="border rounded-lg p-4 overflow-y-auto h-full">
+              <h3 className="text-base font-semibold border-b pb-1 mb-2">Additional Settings</h3>
+              <p className="text-base text-muted-foreground">Additional chart settings can be configured here.</p>
             </div>
           </div>
-        </DialogHeader>
-
-        <div className="grid grid-cols-2 gap-4 flex-1 min-h-0">
-          {/* Left Side - Parameters and DataSource */}
-          <div className="border rounded-lg p-4 overflow-hidden h-full flex flex-col">
-            {/* Tab Navigation */}
-            <div className="flex gap-2 mb-4">
-              <button
-                className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
-                  activeTab === "datasource"
-                    ? "bg-primary text-primary-foreground"
-                    : "bg-muted hover:bg-muted/80"
-                }`}
-                onClick={() => setActiveTab("datasource")}
-              >
-                DataSource
-              </button>
-              <button
-                className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
-                  activeTab === "parameters"
-                    ? "bg-primary text-primary-foreground"
-                    : "bg-muted hover:bg-muted/80"
-                }`}
-                onClick={() => setActiveTab("parameters")}
-              >
-                Parameters
-              </button>
-              <button
-                className={`px-3 py-1 text-sm font-medium rounded-md transition-colors ${
-                  activeTab === "appearance"
-                    ? "bg-primary text-primary-foreground"
-                    : "bg-muted hover:bg-muted/80"
-                }`}
-                onClick={() => setActiveTab("appearance")}
-              >
-                Appearance
-              </button>
-            </div>
-
-            {/* Tab Content */}
-            <div className="flex-1 overflow-y-auto">
-              {/* DataSource Section */}
-              {activeTab === "datasource" && (
-                <div className="space-y-4">
-                  {/* Selected Data Source Section */}
-                  <div className="border rounded-lg p-3 bg-muted/30">
-                    <div className="flex justify-between items-center mb-2">
-                      <h4 className="text-sm font-medium">Selected Data Source</h4>
-                      <div className="flex gap-2">
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          className="h-7 text-xs"
-                          onClick={manualEntry.openForNew}
-                        >
-                          <Plus className="h-3 w-3 mr-1" />
-                          Add Manual Entry
-                        </Button>
-                        <Button
-                          variant="outline"
-                          size="sm"
-                          className="h-7 text-xs"
-                          onClick={() => setTriggerSignalDialogOpen(true)}
-                        >
-                          <Plus className="h-3 w-3 mr-1" />
-                          Signal Search
-                        </Button>
-                      </div>
-                    </div>
-                    
-                    {selectedDataSourceItems.length > 0 ? (
-                      <div className="space-y-3">
-                        <div className="border rounded-lg overflow-hidden">
-                          <Table>
-                            <TableHeader>
-                              <TableRow>
-                                <TableHead className="h-8 text-xs px-2">Plant</TableHead>
-                                <TableHead className="h-8 text-xs px-2">Machine No</TableHead>
-                                <TableHead className="h-8 text-xs px-2">Legend</TableHead>
-                                <TableHead className="h-8 text-xs px-2">Start</TableHead>
-                                <TableHead className="h-8 text-xs px-2">End</TableHead>
-                                <TableHead className="h-8 text-xs w-8"></TableHead>
-                              </TableRow>
-                            </TableHeader>
-                            <TableBody>
-                              {selectedDataSourceItems.map((item) => (
-                                <TableRow key={item.id}>
-                                  <TableCell className="px-2 py-1 text-xs">{item.plant}</TableCell>
-                                  <TableCell className="px-2 py-1 text-xs">{item.machineNo}</TableCell>
-                                  <TableCell className="px-2 py-1 text-xs">
-                                    {item.labelDescription ? `${item.label} (${item.labelDescription})` : item.label}
-                                  </TableCell>
-                                  <TableCell className="px-2 py-1 text-xs">
-                                    <div>
-                                      <div>{item.start.split("T")[0]}</div>
-                                      <div>{item.start.split("T")[1]}</div>
-                                    </div>
-                                  </TableCell>
-                                  <TableCell className="px-2 py-1 text-xs">
-                                    <div>
-                                      <div>{item.end.split("T")[0]}</div>
-                                      <div>{item.end.split("T")[1]}</div>
-                                    </div>
-                                  </TableCell>
-                                  <TableCell className="px-1 py-1">
-                                    <div className="flex gap-1">
-                                      <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className="h-6 w-6 p-0"
-                                        onClick={() => manualEntry.openForEdit(item)}
-                                      >
-                                        <Settings className="h-3 w-3" />
-                                      </Button>
-                                      <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        className="h-6 w-6 p-0"
-                                        onClick={() => {
-                                          setSelectedDataSourceItems(
-                                            selectedDataSourceItems.filter(i => i.id !== item.id)
-                                          )
-                                        }}
-                                      >
-                                        <X className="h-3 w-3" />
-                                      </Button>
-                                    </div>
-                                  </TableCell>
-                                </TableRow>
-                              ))}
-                            </TableBody>
-                          </Table>
-                        </div>
-                        
-                        {/* Time Offset Controls - Collapsible */}
-                        <Collapsible open={offsetSectionOpen} onOpenChange={setOffsetSectionOpen}>
-                          <div className="border rounded-lg p-3 bg-background">
-                            <CollapsibleTrigger asChild>
-                              <div className="flex items-center justify-between cursor-pointer">
-                                <div className="flex items-center gap-2">
-                                  <h4 className="text-sm font-medium">Time Offset</h4>
-                                  {!offsetSectionOpen && (startOffset !== 0 || endOffset !== 0) && (
-                                    <div className="text-xs text-muted-foreground">
-                                      {startOffset !== 0 && `Start: ${startOffset > 0 ? '+' : ''}${startOffset}${startOffsetUnit}`}
-                                      {startOffset !== 0 && endOffset !== 0 && ', '}
-                                      {endOffset !== 0 && `End: ${endOffset > 0 ? '+' : ''}${endOffset}${endOffsetUnit}`}
-                                    </div>
-                                  )}
-                                </div>
-                                <ChevronDown className={`h-4 w-4 transition-transform ${offsetSectionOpen ? 'rotate-180' : ''}`} />
-                              </div>
-                            </CollapsibleTrigger>
-                            
-                            <CollapsibleContent className="overflow-hidden">
-                              <div className="grid grid-cols-2 gap-4 mt-3">
-                                {/* Start Offset */}
-                                <div className="space-y-2">
-                                  <Label className="text-xs font-medium">Start Offset:</Label>
-                                  <div className="flex items-center gap-2">
-                                    <Input
-                                      type="number"
-                                      value={startOffset}
-                                      onChange={(e) => setStartOffset(Number(e.target.value))}
-                                      className="w-16 h-7 text-xs"
-                                      placeholder="0"
-                                    />
-                                    <Select value={startOffsetUnit} onValueChange={(value: 'min' | 'sec') => setStartOffsetUnit(value)}>
-                                      <SelectTrigger className="w-16 h-7 text-xs">
-                                        <SelectValue />
-                                      </SelectTrigger>
-                                      <SelectContent>
-                                        <SelectItem value="min">min</SelectItem>
-                                        <SelectItem value="sec">sec</SelectItem>
-                                      </SelectContent>
-                                    </Select>
-                                    {startOffset !== 0 && (
-                                      <span className="text-xs text-muted-foreground">
-                                        ({startOffset > 0 ? '+' : ''}{startOffset}{startOffsetUnit})
-                                      </span>
-                                    )}
-                                    <Button
-                                      variant="outline"
-                                      size="sm"
-                                      className="h-7 text-xs px-2"
-                                      onClick={resetStartOffset}
-                                      disabled={startOffset === 0}
-                                    >
-                                      Reset
-                                    </Button>
-                                  </div>
-                                </div>
-                                
-                                {/* End Offset */}
-                                <div className="space-y-2">
-                                  <Label className="text-xs font-medium">End Offset:</Label>
-                                  <div className="flex items-center gap-2">
-                                    <Input
-                                      type="number"
-                                      value={endOffset}
-                                      onChange={(e) => setEndOffset(Number(e.target.value))}
-                                      className="w-16 h-7 text-xs"
-                                      placeholder="0"
-                                    />
-                                    <Select value={endOffsetUnit} onValueChange={(value: 'min' | 'sec') => setEndOffsetUnit(value)}>
-                                      <SelectTrigger className="w-16 h-7 text-xs">
-                                        <SelectValue />
-                                      </SelectTrigger>
-                                      <SelectContent>
-                                        <SelectItem value="min">min</SelectItem>
-                                        <SelectItem value="sec">sec</SelectItem>
-                                      </SelectContent>
-                                    </Select>
-                                    {endOffset !== 0 && (
-                                      <span className="text-xs text-muted-foreground">
-                                        ({endOffset > 0 ? '+' : ''}{endOffset}{endOffsetUnit})
-                                      </span>
-                                    )}
-                                    <Button
-                                      variant="outline"
-                                      size="sm"
-                                      className="h-7 text-xs px-2"
-                                      onClick={resetEndOffset}
-                                      disabled={endOffset === 0}
-                                    >
-                                      Reset
-                                    </Button>
-                                  </div>
-                                </div>
-                              </div>
-                            </CollapsibleContent>
-                          </div>
-                        </Collapsible>
-                      </div>
-                    ) : (
-                      <p className="text-xs text-muted-foreground">No data source items selected. Select items from the table below and click 'Add Selected'.</p>
-                    )}
-                  </div>
-
-                  {/* Event Information Section */}
-                  <div>
-                    <div className="flex justify-between items-center mb-3">
-                      <h4 className="text-sm font-medium">Event Information</h4>
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        className="h-7 text-xs"
-                        disabled={selectedEventIds.size === 0}
-                        onClick={() => {
-                          const selectedEvents = events.filter(event => 
-                            selectedEventIds.has(event.id)
-                          )
-                          const newItems = [...selectedDataSourceItems]
-                          selectedEvents.forEach(event => {
-                            if (!newItems.find(item => item.id === event.id)) {
-                              newItems.push(event)
-                            }
-                          })
-                          setSelectedDataSourceItems(newItems)
-                          setSelectedEventIds(new Set())
-                        }}
-                      >
-                        <Plus className="h-3 w-3 mr-1" />
-                        Add Selected ({selectedEventIds.size})
-                      </Button>
-                    </div>
-                    
-                    {/* Search Input */}
-                    <div className="relative mb-3 px-1">
-                      <Search className="absolute left-4 top-1/2 transform -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-                      <Input
-                        placeholder="Search events by plant, machine, label, or event..."
-                        value={eventSearchTerm}
-                        onChange={(e) => setEventSearchTerm(e.target.value)}
-                        className="h-8 text-sm pl-10"
-                      />
-                    </div>
-                    
-                    <div className="border rounded-lg overflow-hidden">
-                      <Table>
-                        <TableHeader>
-                          <TableRow>
-                            <TableHead className="h-8 text-xs w-6 px-1">
-                              <Checkbox
-                                checked={(() => {
-                                  const filteredEvents = events.filter((event) => {
-                                    if (!eventSearchTerm) return true
-                                    const searchLower = eventSearchTerm.toLowerCase()
-                                    return (
-                                      event.plant.toLowerCase().includes(searchLower) ||
-                                      event.machineNo.toLowerCase().includes(searchLower) ||
-                                      event.label.toLowerCase().includes(searchLower) ||
-                                      event.labelDescription?.toLowerCase().includes(searchLower) ||
-                                      event.event.toLowerCase().includes(searchLower) ||
-                                      event.eventDetail?.toLowerCase().includes(searchLower)
-                                    )
-                                  })
-                                  return filteredEvents.length > 0 && filteredEvents.every(event => selectedEventIds.has(event.id))
-                                })()}
-                                onCheckedChange={(checked) => {
-                                  const filteredEvents = events.filter((event) => {
-                                    if (!eventSearchTerm) return true
-                                    const searchLower = eventSearchTerm.toLowerCase()
-                                    return (
-                                      event.plant.toLowerCase().includes(searchLower) ||
-                                      event.machineNo.toLowerCase().includes(searchLower) ||
-                                      event.label.toLowerCase().includes(searchLower) ||
-                                      event.labelDescription?.toLowerCase().includes(searchLower) ||
-                                      event.event.toLowerCase().includes(searchLower) ||
-                                      event.eventDetail?.toLowerCase().includes(searchLower)
-                                    )
-                                  })
-                                  if (checked) {
-                                    const newSelected = new Set(selectedEventIds)
-                                    filteredEvents.forEach(event => newSelected.add(event.id))
-                                    setSelectedEventIds(newSelected)
-                                  } else {
-                                    const newSelected = new Set(selectedEventIds)
-                                    filteredEvents.forEach(event => newSelected.delete(event.id))
-                                    setSelectedEventIds(newSelected)
-                                  }
-                                }}
-                                className="h-3 w-3"
-                              />
-                            </TableHead>
-                            <TableHead className="h-8 text-xs px-2">Plant</TableHead>
-                            <TableHead className="h-8 text-xs px-2">Machine</TableHead>
-                            <TableHead className="h-8 text-xs px-2">Label</TableHead>
-                            <TableHead className="h-8 text-xs px-2">Event</TableHead>
-                          </TableRow>
-                        </TableHeader>
-                        <TableBody>
-                          {events
-                            .filter((event) => {
-                              if (!eventSearchTerm) return true
-                              const searchLower = eventSearchTerm.toLowerCase()
-                              return (
-                                event.plant.toLowerCase().includes(searchLower) ||
-                                event.machineNo.toLowerCase().includes(searchLower) ||
-                                event.label.toLowerCase().includes(searchLower) ||
-                                event.labelDescription?.toLowerCase().includes(searchLower) ||
-                                event.event.toLowerCase().includes(searchLower) ||
-                                event.eventDetail?.toLowerCase().includes(searchLower)
-                              )
-                            })
-                            .map((event) => {
-                            const isInSelectedDataSource = selectedDataSourceItems.some(item => item.id === event.id)
-                            return (
-                              <TableRow 
-                                key={event.id}
-                                className={`cursor-pointer ${selectedEventIds.has(event.id) ? "bg-muted/50" : ""} ${isInSelectedDataSource ? "opacity-50" : ""}`}
-                                title={`${event.event}\nStart: ${event.start.replace("T", " ")}\nEnd: ${event.end.replace("T", " ")}`}
-                              >
-                              <TableCell className="px-1 py-1">
-                                <Checkbox
-                                  checked={selectedEventIds.has(event.id)}
-                                  disabled={isInSelectedDataSource}
-                                  onCheckedChange={(checked) => {
-                                    const newSelectedIds = new Set(selectedEventIds)
-                                    if (checked) {
-                                      newSelectedIds.add(event.id)
-                                    } else {
-                                      newSelectedIds.delete(event.id)
-                                    }
-                                    setSelectedEventIds(newSelectedIds)
-                                  }}
-                                  className="h-3 w-3"
-                                />
-                              </TableCell>
-                              <TableCell className="px-2 py-1 text-xs">{event.plant}</TableCell>
-                              <TableCell className="px-2 py-1 text-xs">{event.machineNo}</TableCell>
-                              <TableCell className="px-2 py-1 text-xs">
-                                <div className="leading-tight">
-                                  <div>{event.label}</div>
-                                  <div className="text-muted-foreground">{event.labelDescription || ""}</div>
-                                </div>
-                              </TableCell>
-                              <TableCell className="px-2 py-1 text-xs">
-                                <div className="leading-tight">
-                                  <div className="flex items-center gap-1">
-                                    {event.event}
-                                    {isInSelectedDataSource && (
-                                      <Badge variant="secondary" className="h-4 text-[10px] px-1">Added</Badge>
-                                    )}
-                                  </div>
-                                  <div className="text-muted-foreground">{event.eventDetail || ""}</div>
-                                </div>
-                              </TableCell>
-                            </TableRow>
-                            )
-                          })}
-                        </TableBody>
-                      </Table>
-                    </div>
-                  </div>
-                </div>
-              )}
-
-              {/* Parameters Section */}
-              {activeTab === "parameters" && (
-                <div className="flex flex-col space-y-4 h-full">
-                  {/* X Parameter Settings */}
-                  <div className="border rounded-lg p-3 bg-muted/30">
-                    <h4 className="font-medium text-sm mb-2">X Parameter Settings</h4>
-                    <div className="flex gap-2">
-                      {/* Parameter Type */}
-                      <div className="w-38">
-                        <Label htmlFor="x-axis-type" className="text-sm mb-1 block">Parameter Type</Label>
-                        <select
-                          id="x-axis-type"
-                          className="w-full h-8 px-2 py-1 border rounded-md text-sm"
-                          value={editingChart.xAxisType || "datetime"}
-                          onChange={(e) => {
-                            setEditingChart({
-                              ...editingChart,
-                              xAxisType: e.target.value as "datetime" | "time" | "parameter",
-                              ...(e.target.value !== "parameter" && { xParameter: "" }),
-                            })
-                          }}
-                        >
-                          <option value="datetime">Datetime</option>
-                          <option value="time">Time (elapsed)</option>
-                          <option value="parameter">Parameter</option>
-                        </select>
-                      </div>
-
-                      {/* Parameter */}
-                      <div className="flex-1">
-                        <Label htmlFor="x-parameter" className="text-sm mb-1 block">Parameter</Label>
-                        <Input
-                          id="x-parameter"
-                          value={editingChart.xParameter || ""}
-                          onChange={(e) => {
-                            setEditingChart({
-                              ...editingChart,
-                              xParameter: e.target.value,
-                            })
-                          }}
-                          placeholder="Enter parameter"
-                          disabled={editingChart.xAxisType !== "parameter"}
-                          className="h-8 text-sm"
-                        />
-                      </div>
-
-                    </div>
-                  </div>
-
-                  {/* Y Parameters Settings */}
-                  <div className="flex flex-col border rounded-lg p-3 bg-muted/30 min-h-0 flex-1">
-                    <div className="flex justify-between items-center mb-2 flex-shrink-0">
-                      <h4 className="font-medium text-sm">Y Parameters Settings</h4>
-                      <Button
-                        variant="outline"
-                        size="sm"
-                        className="h-7 text-xs"
-                        onClick={() => {
-                          const newParam = {
-                            parameterType: "Parameter" as "Parameter" | "Formula",
-                            parameter: "",
-                            axisNo: 1,
-                            axisName: "",
-                            range: {
-                              auto: true,
-                              min: 0,
-                              max: 100
-                            },
-                            line: {
-                              width: 2,
-                              color: "#000000",
-                              style: "solid" as const
-                            }
-                          }
-                          const newParams = [...(editingChart.yAxisParams || []), newParam]
-                          setEditingChart({
-                            ...editingChart,
-                            yAxisParams: newParams
-                          })
-                          setLastAddedParamIndex(newParams.length - 1)
-                        }}
-                      >
-                        Add Y Parameter
-                      </Button>
-                    </div>
-                    
-                    {/* Table Header */}
-                    <div className="flex gap-2 mb-2 px-1 pb-1 border-b flex-shrink-0">
-                      <div className="w-28 text-xs font-medium text-muted-foreground">Parameter Type</div>
-                      <div className="flex-1 text-xs font-medium text-muted-foreground">Parameter</div>
-                      <div className="w-16 text-xs font-medium text-muted-foreground">Axis No</div>
-                      <div className="w-7"></div>
-                    </div>
-                    
-                    {/* Y Parameter List */}
-                    <div className="flex-1 overflow-y-auto min-h-0">
-                      <div className="space-y-1">
-                        {editingChart.yAxisParams?.map((param, index) => (
-                          <div key={index} className="flex gap-2 p-1">
-                            {/* Parameter Type */}
-                            <div className="w-28">
-                              <select
-                                value={param.parameterType || "Parameter"}
-                                onChange={(e) => {
-                                  const newParams = [...(editingChart.yAxisParams || [])]
-                                  newParams[index] = { ...newParams[index], parameterType: e.target.value as "Parameter" | "Formula" }
-                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                }}
-                                className="w-full h-7 px-2 py-1 border rounded-md text-sm"
-                              >
-                                <option value="Parameter">Parameter</option>
-                                <option value="Formula">Formula</option>
-                              </select>
-                            </div>
-                            
-                            {/* Parameter */}
-                            <div className="flex-1">
-                              <Input
-                                ref={(el) => {
-                                  parameterInputRefs.current[index] = el
-                                }}
-                                value={param.parameter}
-                                onChange={(e) => {
-                                  const newParams = [...(editingChart.yAxisParams || [])]
-                                  newParams[index] = { ...newParams[index], parameter: e.target.value }
-                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                }}
-                                placeholder="Parameter"
-                                className="h-7 text-sm"
-                              />
-                            </div>
-                            
-                            {/* Axis No */}
-                            <div className="w-16">
-                              <Input
-                                type="number"
-                                min="1"
-                                max="10"
-                                value={param.axisNo || 1}
-                                onChange={(e) => {
-                                  const newParams = [...(editingChart.yAxisParams || [])]
-                                  newParams[index] = { ...newParams[index], axisNo: parseInt(e.target.value) || 1 }
-                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                }}
-                                className="h-7 text-sm"
-                              />
-                            </div>
-                            
-                            {/* Delete Button */}
-                            <div className="w-7">
-                              <Button
-                                variant="ghost"
-                                size="sm"
-                                className="h-7 w-7 p-0"
-                                onClick={() => {
-                                  const newParams = editingChart.yAxisParams?.filter((_, i) => i !== index) || []
-                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                }}
-                              >
-                                <X className="h-3 w-3" />
-                              </Button>
-                            </div>
-                          </div>
-                        )) || <p className="text-sm text-muted-foreground px-1">No Y parameters added yet.</p>}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              )}
-
-              {/* Appearance Section */}
-              {activeTab === "appearance" && (
-                <div className="space-y-4">
-                  <div>
-                    <Label htmlFor="chart-title" className="text-sm">Title</Label>
-                    <div className="flex items-center gap-2">
-                      <Input
-                        id="chart-title"
-                        value={editingChart.title}
-                        onChange={(e) => {
-                          setEditingChart({
-                            ...editingChart,
-                            title: e.target.value,
-                          })
-                        }}
-                        className="h-8 text-sm flex-1"
-                      />
-                      <div className="flex items-center space-x-4">
-                        <div className="flex items-center space-x-2">
-                          <input
-                            type="checkbox"
-                            id="show-title"
-                            checked={editingChart.showTitle ?? true}
-                            onChange={(e) => {
-                              setEditingChart({
-                                ...editingChart,
-                                showTitle: e.target.checked,
-                              })
-                            }}
-                            className="rounded"
-                          />
-                          <Label htmlFor="show-title" className="text-sm whitespace-nowrap">Show Title</Label>
-                        </div>
-                        <div className="flex items-center space-x-2">
-                          <input
-                            type="checkbox"
-                            id="show-legend"
-                            checked={editingChart.legend ?? true}
-                            onChange={(e) => {
-                              setEditingChart({
-                                ...editingChart,
-                                legend: e.target.checked,
-                              })
-                            }}
-                            className="rounded"
-                          />
-                          <Label htmlFor="show-legend" className="text-sm whitespace-nowrap">Show Legend</Label>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-
-                  <div>
-                    <Label htmlFor="x-label" className="text-sm">X-axis Label</Label>
-                    <div className="flex items-center gap-2">
-                      <Input
-                        id="x-label"
-                        value={editingChart.xLabel || ""}
-                        onChange={(e) => {
-                          setEditingChart({
-                            ...editingChart,
-                            xLabel: e.target.value,
-                          })
-                        }}
-                        placeholder="X-axis label"
-                        className="h-8 text-sm flex-1"
-                      />
-                      <div className="min-w-0">
-                        <Popover>
-                          <PopoverTrigger asChild>
-                            <Button variant="outline" size="sm" className="h-8 px-2 text-sm whitespace-nowrap">
-                              <Settings className="h-3 w-3 mr-1" />
-                              Range: {editingChart.xAxisRange?.auto !== false ? 
-                                "Auto" : 
-                                `${editingChart.xAxisRange?.min || "0"} - ${editingChart.xAxisRange?.max || "100"}`
-                              }
-                            </Button>
-                          </PopoverTrigger>
-                          <PopoverContent className="w-80" align="end">
-                            <div className="space-y-4">
-                              <h4 className="font-medium text-sm">X Range Settings</h4>
-                              
-                              <div className="flex items-center gap-2">
-                                <input
-                                  type="checkbox"
-                                  id="x-range-auto"
-                                  checked={editingChart.xAxisRange?.auto ?? true}
-                                  onChange={(e) => {
-                                    setEditingChart({
-                                      ...editingChart,
-                                      xAxisRange: {
-                                        ...editingChart.xAxisRange,
-                                        auto: e.target.checked,
-                                        min: editingChart.xAxisRange?.min || "",
-                                        max: editingChart.xAxisRange?.max || ""
-                                      }
-                                    })
-                                  }}
-                                  className="rounded"
-                                />
-                                <Label htmlFor="x-range-auto" className="text-sm">Auto Range</Label>
-                              </div>
-                              
-                              {(() => {
-                                const axisType = editingChart.xAxisType || "datetime"
-                                
-                                if (axisType === "datetime") {
-                                  return (
-                                    <div className="space-y-3">
-                                      <div>
-                                        <Label htmlFor="x-range-min" className="text-sm">Start Date/Time</Label>
-                                        <Input
-                                          id="x-range-min"
-                                          type="datetime-local"
-                                          value={editingChart.xAxisRange?.min || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: e.target.value,
-                                                max: editingChart.xAxisRange?.max || ""
-                                              }
-                                            })
-                                          }}
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                      <div>
-                                        <Label htmlFor="x-range-max" className="text-sm">End Date/Time</Label>
-                                        <Input
-                                          id="x-range-max"
-                                          type="datetime-local"
-                                          value={editingChart.xAxisRange?.max || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: editingChart.xAxisRange?.min || "",
-                                                max: e.target.value
-                                              }
-                                            })
-                                          }}
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                    </div>
-                                  )
-                                } else if (axisType === "time") {
-                                  return (
-                                    <div className="space-y-3">
-                                      <div>
-                                        <Label className="text-sm">Time Unit</Label>
-                                        <select
-                                          className="w-full h-8 px-2 py-1 border rounded-md text-sm"
-                                          value={editingChart.xAxisRange?.unit || "sec"}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                unit: e.target.value as "sec" | "min" | "hr",
-                                                min: editingChart.xAxisRange?.min ?? 0,
-                                                max: editingChart.xAxisRange?.max ?? 100
-                                              }
-                                            })
-                                          }}
-                                        >
-                                          <option value="sec">Seconds</option>
-                                          <option value="min">Minutes</option>
-                                          <option value="hr">Hours</option>
-                                        </select>
-                                      </div>
-                                      <div>
-                                        <Label htmlFor="x-range-min" className="text-sm">Start Time</Label>
-                                        <Input
-                                          id="x-range-min"
-                                          type="number"
-                                          value={editingChart.xAxisRange?.min || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: e.target.value,
-                                                max: editingChart.xAxisRange?.max || ""
-                                              }
-                                            })
-                                          }}
-                                          placeholder="0"
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                      <div>
-                                        <Label htmlFor="x-range-max" className="text-sm">End Time</Label>
-                                        <Input
-                                          id="x-range-max"
-                                          type="number"
-                                          value={editingChart.xAxisRange?.max || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: editingChart.xAxisRange?.min || "",
-                                                max: e.target.value
-                                              }
-                                            })
-                                          }}
-                                          placeholder="100"
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                    </div>
-                                  )
-                                } else {
-                                  // Parameter type
-                                  return (
-                                    <div className="space-y-3">
-                                      <div>
-                                        <Label htmlFor="x-range-min" className="text-sm">Min Value</Label>
-                                        <Input
-                                          id="x-range-min"
-                                          type="number"
-                                          value={editingChart.xAxisRange?.min || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: e.target.value,
-                                                max: editingChart.xAxisRange?.max || ""
-                                              }
-                                            })
-                                          }}
-                                          placeholder="Enter min value"
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                      <div>
-                                        <Label htmlFor="x-range-max" className="text-sm">Max Value</Label>
-                                        <Input
-                                          id="x-range-max"
-                                          type="number"
-                                          value={editingChart.xAxisRange?.max || ""}
-                                          onChange={(e) => {
-                                            setEditingChart({
-                                              ...editingChart,
-                                              xAxisRange: {
-                                                ...editingChart.xAxisRange,
-                                                auto: false,
-                                                min: editingChart.xAxisRange?.min || "",
-                                                max: e.target.value
-                                              }
-                                            })
-                                          }}
-                                          placeholder="Enter max value"
-                                          disabled={editingChart.xAxisRange?.auto ?? true}
-                                          className="h-8 text-sm"
-                                        />
-                                      </div>
-                                    </div>
-                                  )
-                                }
-                              })()}
-                            </div>
-                          </PopoverContent>
-                        </Popover>
-                      </div>
-                    </div>
-                  </div>
-
-                  <div>
-                    <Label className="text-sm">Y-axis Labels</Label>
-                    <div className="mt-1 max-h-24 overflow-y-auto border rounded-md p-2">
-                      <div className="space-y-2">
-                        {(() => {
-                          // Check if Y parameters exist
-                          if (!editingChart.yAxisParams || editingChart.yAxisParams.length === 0) {
-                            return null
-                          }
-                          
-                          // Get unique axis numbers from Y parameters
-                          const axisNumbers = [...new Set(
-                            editingChart.yAxisParams.map(param => param.axisNo || 1)
-                          )].sort((a, b) => a - b)
-                          
-                          return axisNumbers.map(axisNo => {
-                            // Find the first parameter with this axis number to get its range
-                            const axisParam = editingChart.yAxisParams?.find(param => param.axisNo === axisNo)
-                            
-                            return (
-                              <div key={axisNo} className="flex items-center gap-2">
-                                <Label className="text-xs w-16 text-muted-foreground">Axis {axisNo}:</Label>
-                                <Input
-                                  value={editingChart.yAxisLabels?.[axisNo] || ""}
-                                  onChange={(e) => {
-                                    setEditingChart({
-                                      ...editingChart,
-                                      yAxisLabels: {
-                                        ...editingChart.yAxisLabels,
-                                        [axisNo]: e.target.value,
-                                      },
-                                    })
-                                  }}
-                                  placeholder={`Y-axis ${axisNo} label`}
-                                  className="h-8 text-sm flex-1"
-                                />
-                                <div className="min-w-0">
-                                  <Popover>
-                                    <PopoverTrigger asChild>
-                                      <Button variant="outline" size="sm" className="h-8 px-2 text-sm whitespace-nowrap">
-                                        <Settings className="h-3 w-3 mr-1" />
-                                        Range: {axisParam?.range?.auto !== false ? 
-                                          "Auto" : 
-                                          `${axisParam?.range?.min || 0} - ${axisParam?.range?.max || 100}`
-                                        }
-                                      </Button>
-                                    </PopoverTrigger>
-                                    <PopoverContent className="w-80" align="end">
-                                      <div className="space-y-4">
-                                        <h4 className="font-medium text-sm">Y Range Settings (Axis {axisNo})</h4>
-                                        
-                                        <div className="flex items-center gap-2">
-                                          <input
-                                            type="checkbox"
-                                            id={`y-range-auto-${axisNo}`}
-                                            checked={axisParam?.range?.auto ?? true}
-                                            onChange={(e) => {
-                                              const newParams = [...(editingChart.yAxisParams || [])]
-                                              const paramIndex = newParams.findIndex(param => param.axisNo === axisNo)
-                                              if (paramIndex >= 0) {
-                                                newParams[paramIndex] = {
-                                                  ...newParams[paramIndex],
-                                                  range: {
-                                                    ...newParams[paramIndex].range,
-                                                    auto: e.target.checked,
-                                                    min: newParams[paramIndex].range?.min || 0,
-                                                    max: newParams[paramIndex].range?.max || 100
-                                                  }
-                                                }
-                                                setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                              }
-                                            }}
-                                            className="rounded"
-                                          />
-                                          <Label htmlFor={`y-range-auto-${axisNo}`} className="text-sm">Auto Range</Label>
-                                        </div>
-                                        
-                                        <div className="space-y-3">
-                                          <div>
-                                            <Label htmlFor={`y-range-min-${axisNo}`} className="text-sm">Min Value</Label>
-                                            <Input
-                                              id={`y-range-min-${axisNo}`}
-                                              type="number"
-                                              value={axisParam?.range?.min || 0}
-                                              onChange={(e) => {
-                                                const newParams = [...(editingChart.yAxisParams || [])]
-                                                const paramIndex = newParams.findIndex(param => param.axisNo === axisNo)
-                                                if (paramIndex >= 0) {
-                                                  newParams[paramIndex] = {
-                                                    ...newParams[paramIndex],
-                                                    range: {
-                                                      ...newParams[paramIndex].range,
-                                                      auto: false,
-                                                      min: parseFloat(e.target.value) || 0,
-                                                      max: newParams[paramIndex].range?.max || 100
-                                                    }
-                                                  }
-                                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                }
-                                              }}
-                                              placeholder="Enter min value"
-                                              disabled={axisParam?.range?.auto ?? true}
-                                              className="h-8 text-sm"
-                                            />
-                                          </div>
-                                          <div>
-                                            <Label htmlFor={`y-range-max-${axisNo}`} className="text-sm">Max Value</Label>
-                                            <Input
-                                              id={`y-range-max-${axisNo}`}
-                                              type="number"
-                                              value={axisParam?.range?.max || 100}
-                                              onChange={(e) => {
-                                                const newParams = [...(editingChart.yAxisParams || [])]
-                                                const paramIndex = newParams.findIndex(param => param.axisNo === axisNo)
-                                                if (paramIndex >= 0) {
-                                                  newParams[paramIndex] = {
-                                                    ...newParams[paramIndex],
-                                                    range: {
-                                                      ...newParams[paramIndex].range,
-                                                      auto: false,
-                                                      min: newParams[paramIndex].range?.min || 0,
-                                                      max: parseFloat(e.target.value) || 100
-                                                    }
-                                                  }
-                                                  setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                }
-                                              }}
-                                              placeholder="Enter max value"
-                                              disabled={axisParam?.range?.auto ?? true}
-                                              className="h-8 text-sm"
-                                            />
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </PopoverContent>
-                                  </Popover>
-                                </div>
-                              </div>
-                            )
-                          })
-                        })()} 
-                        {(!editingChart.yAxisParams || editingChart.yAxisParams.length === 0) && (
-                          <p className="text-xs text-muted-foreground">Add Y parameters to configure axis labels</p>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-
-                  {/* Appearance Settings Table */}
-                  <div>
-                    <Label className="text-sm mb-2 block">Appearance Settings</Label>
-                    <div className="border rounded-lg overflow-hidden">
-                      <Table>
-                        <TableHeader>
-                          <TableRow>
-                            <TableHead className="text-xs">Data Source</TableHead>
-                            <TableHead className="text-xs">Parameter</TableHead>
-                            <TableHead className="text-xs">Marker</TableHead>
-                            <TableHead className="text-xs">Line</TableHead>
-                          </TableRow>
-                        </TableHeader>
-                        <TableBody>
-                          {selectedDataSourceItems.length > 0 && editingChart.yAxisParams && editingChart.yAxisParams.length > 0 ? (
-                            selectedDataSourceItems.flatMap((dataSource) =>
-                              editingChart.yAxisParams?.map((param, paramIndex) => {
-                                const key = `${dataSource.id}-${paramIndex}`
-                                return (
-                                  <TableRow key={key}>
-                                    <TableCell className="text-xs">
-                                      <div>
-                                        <div className="font-medium">{dataSource.plant} - {dataSource.machineNo}</div>
-                                        <div className="text-muted-foreground">
-                                          {dataSource.labelDescription ? `${dataSource.label} (${dataSource.labelDescription})` : dataSource.label}
-                                        </div>
-                                      </div>
-                                    </TableCell>
-                                    <TableCell className="text-xs">
-                                      <div>
-                                        <div className="font-medium flex items-center gap-1">
-                                          {param.parameter || "Unnamed"}
-                                          {param.parameterType === "Formula" && (
-                                            <Badge variant="secondary" className="text-[10px] px-1 h-4">Formula</Badge>
-                                          )}
-                                        </div>
-                                        <div className="text-muted-foreground">Axis {param.axisNo || 1}</div>
-                                      </div>
-                                    </TableCell>
-                                    <TableCell className="text-xs">
-                                      <div className="flex items-center gap-1">
-                                        <Popover>
-                                          <PopoverTrigger asChild>
-                                            <Button variant="outline" size="sm" className="h-7 text-xs">
-                                              <div className="flex items-center gap-1">
-                                                <div 
-                                                  className="w-3 h-3 border rounded-full" 
-                                                  style={{
-                                                    backgroundColor: param.marker?.fillColor || "#0066cc",
-                                                    borderColor: param.marker?.borderColor || "#0066cc"
-                                                  }}
-                                                />
-                                                <span>{param.marker?.type || "circle"}</span>
-                                              </div>
-                                            </Button>
-                                          </PopoverTrigger>
-                                          <PopoverContent className="w-80">
-                                            <div className="space-y-3">
-                                              <h4 className="text-sm font-medium">Marker Settings</h4>
-                                              
-                                              <div>
-                                                <Label className="text-xs">Type</Label>
-                                                <select
-                                                  value={param.marker?.type || "circle"}
-                                                  onChange={(e) => {
-                                                    const newParams = [...(editingChart.yAxisParams || [])]
-                                                    newParams[paramIndex] = {
-                                                      ...newParams[paramIndex],
-                                                      marker: {
-                                                        ...newParams[paramIndex].marker,
-                                                        type: e.target.value as MarkerType,
-                                                        size: newParams[paramIndex].marker?.size || 6,
-                                                        borderColor: newParams[paramIndex].marker?.borderColor || "#0066cc",
-                                                        fillColor: newParams[paramIndex].marker?.fillColor || "#0066cc"
-                                                      }
-                                                    }
-                                                    setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                  }}
-                                                  className="w-full h-7 px-2 py-1 border rounded-md text-xs"
-                                                >
-                                                  <option value="circle">Circle</option>
-                                                  <option value="square">Square</option>
-                                                  <option value="triangle">Triangle</option>
-                                                  <option value="diamond">Diamond</option>
-                                                  <option value="star">Star</option>
-                                                  <option value="cross">Cross</option>
-                                                </select>
-                                              </div>
-                                              
-                                              <div>
-                                                <Label className="text-xs">Size</Label>
-                                                <Input
-                                                  type="number"
-                                                  min="1"
-                                                  max="20"
-                                                  value={param.marker?.size || 6}
-                                                  onChange={(e) => {
-                                                    const newParams = [...(editingChart.yAxisParams || [])]
-                                                    newParams[paramIndex] = {
-                                                      ...newParams[paramIndex],
-                                                      marker: {
-                                                        ...newParams[paramIndex].marker,
-                                                        type: newParams[paramIndex].marker?.type || "circle",
-                                                        size: parseInt(e.target.value) || 6,
-                                                        borderColor: newParams[paramIndex].marker?.borderColor || "#0066cc",
-                                                        fillColor: newParams[paramIndex].marker?.fillColor || "#0066cc"
-                                                      }
-                                                    }
-                                                    setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                  }}
-                                                  className="h-7 text-xs"
-                                                />
-                                              </div>
-                                              
-                                              <div className="grid grid-cols-2 gap-2">
-                                                <div>
-                                                  <Label className="text-xs">Border Color</Label>
-                                                  <Input
-                                                    type="color"
-                                                    value={param.marker?.borderColor || "#0066cc"}
-                                                    onChange={(e) => {
-                                                      const newParams = [...(editingChart.yAxisParams || [])]
-                                                      newParams[paramIndex] = {
-                                                        ...newParams[paramIndex],
-                                                        marker: {
-                                                          ...newParams[paramIndex].marker,
-                                                          type: newParams[paramIndex].marker?.type || "circle",
-                                                          size: newParams[paramIndex].marker?.size || 6,
-                                                          borderColor: e.target.value,
-                                                          fillColor: newParams[paramIndex].marker?.fillColor || "#0066cc"
-                                                        }
-                                                      }
-                                                      setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                    }}
-                                                    className="h-7 w-full"
-                                                  />
-                                                </div>
-                                                <div>
-                                                  <Label className="text-xs">Fill Color</Label>
-                                                  <Input
-                                                    type="color"
-                                                    value={param.marker?.fillColor || "#0066cc"}
-                                                    onChange={(e) => {
-                                                      const newParams = [...(editingChart.yAxisParams || [])]
-                                                      newParams[paramIndex] = {
-                                                        ...newParams[paramIndex],
-                                                        marker: {
-                                                          ...newParams[paramIndex].marker,
-                                                          type: newParams[paramIndex].marker?.type || "circle",
-                                                          size: newParams[paramIndex].marker?.size || 6,
-                                                          borderColor: newParams[paramIndex].marker?.borderColor || "#0066cc",
-                                                          fillColor: e.target.value
-                                                        }
-                                                      }
-                                                      setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                    }}
-                                                    className="h-7 w-full"
-                                                  />
-                                                </div>
-                                              </div>
-                                            </div>
-                                          </PopoverContent>
-                                        </Popover>
-                                      </div>
-                                    </TableCell>
-                                    <TableCell className="text-xs">
-                                      <div className="flex items-center gap-1">
-                                        <Popover>
-                                          <PopoverTrigger asChild>
-                                            <Button variant="outline" size="sm" className="h-7 text-xs">
-                                              <div className="flex items-center gap-1">
-                                                <div 
-                                                  className="w-4 h-0 border-b-2"
-                                                  style={{
-                                                    borderColor: param.line?.color || "#0066cc",
-                                                    borderStyle: param.line?.style || "solid"
-                                                  }}
-                                                />
-                                                <span>{param.line?.style || "solid"}</span>
-                                              </div>
-                                            </Button>
-                                          </PopoverTrigger>
-                                          <PopoverContent className="w-80">
-                                            <div className="space-y-3">
-                                              <h4 className="text-sm font-medium">Line Settings</h4>
-                                              
-                                              <div>
-                                                <Label className="text-xs">Style</Label>
-                                                <select
-                                                  value={param.line?.style || "solid"}
-                                                  onChange={(e) => {
-                                                    const newParams = [...(editingChart.yAxisParams || [])]
-                                                    newParams[paramIndex] = {
-                                                      ...newParams[paramIndex],
-                                                      line: {
-                                                        ...newParams[paramIndex].line,
-                                                        style: e.target.value as LineStyle,
-                                                        width: newParams[paramIndex].line?.width || 2,
-                                                        color: newParams[paramIndex].line?.color || "#0066cc"
-                                                      }
-                                                    }
-                                                    setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                  }}
-                                                  className="w-full h-7 px-2 py-1 border rounded-md text-xs"
-                                                >
-                                                  <option value="solid">Solid</option>
-                                                  <option value="dashed">Dashed</option>
-                                                  <option value="dotted">Dotted</option>
-                                                </select>
-                                              </div>
-                                              
-                                              <div>
-                                                <Label className="text-xs">Width</Label>
-                                                <Input
-                                                  type="number"
-                                                  min="1"
-                                                  max="10"
-                                                  value={param.line?.width || 2}
-                                                  onChange={(e) => {
-                                                    const newParams = [...(editingChart.yAxisParams || [])]
-                                                    newParams[paramIndex] = {
-                                                      ...newParams[paramIndex],
-                                                      line: {
-                                                        ...newParams[paramIndex].line,
-                                                        style: newParams[paramIndex].line?.style || "solid",
-                                                        width: parseInt(e.target.value) || 2,
-                                                        color: newParams[paramIndex].line?.color || "#0066cc"
-                                                      }
-                                                    }
-                                                    setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                  }}
-                                                  className="h-7 text-xs"
-                                                />
-                                              </div>
-                                              
-                                              <div>
-                                                <Label className="text-xs">Color</Label>
-                                                <Input
-                                                  type="color"
-                                                  value={param.line?.color || "#0066cc"}
-                                                  onChange={(e) => {
-                                                    const newParams = [...(editingChart.yAxisParams || [])]
-                                                    newParams[paramIndex] = {
-                                                      ...newParams[paramIndex],
-                                                      line: {
-                                                        ...newParams[paramIndex].line,
-                                                        style: newParams[paramIndex].line?.style || "solid",
-                                                        width: newParams[paramIndex].line?.width || 2,
-                                                        color: e.target.value
-                                                      }
-                                                    }
-                                                    setEditingChart({ ...editingChart, yAxisParams: newParams })
-                                                  }}
-                                                  className="h-7 w-full"
-                                                />
-                                              </div>
-                                            </div>
-                                          </PopoverContent>
-                                        </Popover>
-                                      </div>
-                                    </TableCell>
-                                  </TableRow>
-                                )
-                              }) || []
-                            )
-                          ) : (
-                            <TableRow>
-                              <TableCell colSpan={4} className="text-center text-xs text-muted-foreground py-4">
-                                {selectedDataSourceItems.length === 0 
-                                  ? "No data sources selected. Please select data sources in the DataSource tab."
-                                  : "No Y parameters configured. Please add parameters in the Parameters tab."}
-                              </TableCell>
-                            </TableRow>
-                          )}
-                        </TableBody>
-                      </Table>
-                    </div>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-
-          {/* Right Side - Additional Settings */}
-          <div className="border rounded-lg p-4 overflow-y-auto h-full">
-            <h3 className="text-base font-semibold border-b pb-1 mb-2">Additional Settings</h3>
-            <p className="text-base text-muted-foreground">Additional chart settings can be configured here.</p>
-          </div>
-        </div>
-
-      </DialogContent>
-    </Dialog>
-
-    <ManualEntryDialog
-      isOpen={manualEntry.isOpen}
-      editingItemId={manualEntry.editingItemId}
-      data={manualEntry.data}
-      onClose={manualEntry.close}
-      onUpdateData={manualEntry.updateData}
-      onSave={handleSaveManualEntry}
-      isValid={manualEntry.isValid()}
-    />
-    
-    <TriggerSignalDialog
-      isOpen={triggerSignalDialogOpen}
-      onClose={() => setTriggerSignalDialogOpen(false)}
-      onAddToDataSource={handleAddTriggerSignalResults}
-      availableEvents={events}
-    />
+        </DialogContent>
+      </Dialog>
     </>
   )
 }
+
